refactor feature toggle
